feat: add Velopack auto-update + time-based hatching
- Add Velopack for GitHub Releases auto-update
- Add time-based hatch progress (0.1/sec)
- Fix golden dragon sprite direction (ID 48)
- Add GitHub Actions workflow for automated releases

+++ TypingTamagotchi/Models/Egg.cs
@@ -6,7 +6,7 @@ public class Egg
     public string Name { get; set; } = string.Empty;
     public string SpritePath { get; set; } = string.Empty;
     public int RequiredCount { get; set; }  // 500~2000 랜덤
-    public int CurrentCount { get; set; }
+    public double CurrentCount { get; set; }  // 타이핑 1타 = 1, 시간 초당 = 0.1
 
     public double Progress => RequiredCount > 0
         ? (double)CurrentCount / RequiredCount
